Moved IVersionConverter into its own file
